Refactoring: moved manager classes to management package
