+++ com.apt.project.eshop/src/it/java/com/apt/project/eshop/controller/EShopControllerIT.java
@@ -18,12 +18,12 @@ import org.mockito.MockitoAnnotations;
 import org.mockito.Spy;
 import org.testcontainers.containers.GenericContainer;
 
+import com.apt.project.eshop.management.ShopManager;
+import com.apt.project.eshop.management.TransactionManager;
+import com.apt.project.eshop.management.TransactionalShopManager;
 import com.apt.project.eshop.model.Product;
 import com.apt.project.eshop.repository.CartRepository;
 import com.apt.project.eshop.repository.ProductRepository;
-import com.apt.project.eshop.repository.ShopManager;
-import com.apt.project.eshop.repository.TransactionManager;
-import com.apt.project.eshop.repository.TransactionalShopManager;
 import com.apt.project.eshop.repository.mongo.CartMongoRepository;
 import com.apt.project.eshop.repository.mongo.ProductMongoRepository;
 import com.apt.project.eshop.view.EShopView;
