+++ com.apt.project.eshop/src/it/java/com/apt/project/eshop/view/swing/EShopSwingViewIT.java
@@ -18,11 +18,11 @@ import org.junit.runner.RunWith;
 import org.testcontainers.containers.GenericContainer;
 
 import com.apt.project.eshop.controller.EShopController;
+import com.apt.project.eshop.management.ShopManager;
+import com.apt.project.eshop.management.TransactionManager;
+import com.apt.project.eshop.management.TransactionalShopManager;
 import com.apt.project.eshop.model.Product;
 import com.apt.project.eshop.repository.RepositoryException;
-import com.apt.project.eshop.repository.ShopManager;
-import com.apt.project.eshop.repository.TransactionManager;
-import com.apt.project.eshop.repository.TransactionalShopManager;
 import com.apt.project.eshop.repository.mongo.CartMongoRepository;
 import com.apt.project.eshop.repository.mongo.ProductMongoRepository;
 import com.mongodb.MongoClient;
