+++ com.apt.project.eshop/src/main/java/com/apt/project/eshop/controller/EShopController.java
@@ -2,8 +2,8 @@ package com.apt.project.eshop.controller;
 
 import java.util.List;
 
+import com.apt.project.eshop.management.ShopManager;
 import com.apt.project.eshop.model.Product;
-import com.apt.project.eshop.repository.ShopManager;
 import com.apt.project.eshop.view.EShopView;
 
 public class EShopController {
