+++ com.apt.project.eshop/src/main/java/com/apt/project/eshop/management/ShopManager.java
@@ -1,9 +1,10 @@
-package com.apt.project.eshop.repository;
+package com.apt.project.eshop.management;
 
 import java.util.List;
 
 import com.apt.project.eshop.controller.EShopController;
 import com.apt.project.eshop.model.Product;
+import com.apt.project.eshop.repository.RepositoryException;
 
 public class ShopManager {
