+++ com.apt.project.eshop/src/main/java/com/apt/project/eshop/management/TransactionManager.java
@@ -1,4 +1,4 @@
-package com.apt.project.eshop.repository;
+package com.apt.project.eshop.management;
 
 public interface TransactionManager {
 	<T> T doInTransaction(TransactionCode<T> code);
