+++ com.apt.project.eshop/src/main/java/com/apt/project/eshop/management/TransactionalShopManager.java
@@ -1,4 +1,4 @@
-package com.apt.project.eshop.repository;
+package com.apt.project.eshop.management;
 
 import java.util.logging.Level;
 import java.util.logging.Logger;
