--- com.apt.project.eshop/src/main/java/com/apt/project/eshop/repository/TransactionCode.java
@@ -1,6 +1,0 @@
-package com.apt.project.eshop.repository;
-
-@FunctionalInterface
-public interface TransactionCode<T> {
-	T apply(ProductRepository productRepository, CartRepository cartRepository) throws RepositoryException;
-}
